script: Add backup-dependent-archives.sh and remove some scripts

+++ script/lib/backup.sh
@@ -7,8 +7,13 @@ function backup_github_release() {
     bucket=$4
 
     source_url="https://github.com/${repository}/releases/download/${version}/${name}"
+    key="github.com/${repository}/releases/download/${version}/${name}"
 
-    curl -s -L -O "$source_url"
-    mc cp "${name}" incluster/${bucket}/github.com/${repository}/releases/download/${version}/${name}
-    rm "${name}"
+    if ! mc stat "incluster/${bucket}/${key}" > /dev/null 2>&1; then
+        curl -s -L -O "$source_url"
+        mc cp "${name}" "incluster/${bucket}/${key}"
+        rm "${name}"
+    else
+        echo "Skip to download because ${key} is already exists."
+    fi
 }
